requirements.txt
flask==2.3.3
msal==1.22.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
python-dateutil==2.8.2 